Removed IArticleWordRepository.

+++ src/EFPersistenceModel/ArticleWordRepository.cs
@@ -5,9 +5,10 @@
     using System.Threading.Tasks;
     using ArticleHarbor.DomainModel;
     using ArticleHarbor.EFDataAccess;
+    using Article = DomainModel.Article;
     using ArticleWord = DomainModel.ArticleWord;
 
-    public class ArticleWordRepository : IArticleWordRepository
+    public class ArticleWordRepository : IRepository<ArticleWord>
     {
         private readonly ArticleHarborDbContext context;
 
@@ -33,7 +34,7 @@
             return articleWord == null ? null : articleWord.ToDomain();
         }
 
-        public Task InsertAsync(ArticleWord articleWord)
+        public Task<ArticleWord> InsertAsync(ArticleWord articleWord)
         {
             if (articleWord == null)
                 throw new ArgumentNullException("articleWord");
@@ -41,13 +42,56 @@
             if (this.Select(articleWord.ArticleId, articleWord.Word) == null)
                 this.context.ArticleWords.Add(articleWord.ToPersistence());
 
-            return Task.FromResult<object>(null);
+            return Task.FromResult<ArticleWord>(articleWord);
         }
 
         public Task DeleteAsync(int articleId)
         {
             var articleWords = this.context.ArticleWords
                 .Where(x => x.ArticleId == articleId).ToArray();
+
+            foreach (var articleWord in articleWords)
+                this.context.ArticleWords.Remove(articleWord);
+
+            return Task.FromResult<object>(null);
+        }
+
+        public Task<DomainModel.Article> FineAsync(params object[] identity)
+        {
+            if (identity == null)
+                throw new ArgumentNullException("identity");
+
+            throw new NotImplementedException();
+        }
+
+        public Task<System.Collections.Generic.IEnumerable<ArticleWord>> SelectAsync()
+        {
+            throw new NotImplementedException();
+        }
+
+        Task<Article> IRepository<ArticleWord>.InsertAsync(ArticleWord article)
+        {
+            if (article == null)
+                throw new ArgumentNullException("article");
+
+            throw new NotImplementedException();
+        }
+
+        public Task UpdateAsync(ArticleWord article)
+        {
+            if (article == null)
+                throw new ArgumentNullException("article");
+            throw new NotImplementedException();
+        }
+
+        public Task DeleteAsync(params object[] identity)
+        {
+            if (identity == null)
+                throw new ArgumentNullException("identity");
+
+            var id = (int)identity[0];
+            var articleWords = this.context.ArticleWords
+                .Where(x => x.ArticleId == id).ToArray();
 
             foreach (var articleWord in articleWords)
                 this.context.ArticleWords.Remove(articleWord);
